Trying new approach for finding acq/gui transitions.
Instead of using the cmd_states NPNT state to set the endpoint for the
telemetry retrieve, this new piece uses the aiprops database table
NPNT state.  Also finds the transition points to GUID and then looks
for the max ACQ time before that GUI transition.  The first ACQ/GUI
transitions is used.  The old code failed more often in the case of
two ACQ intervals for the same observation, as it would find the max
ACQ for the obsid after the first GUI transition.