Add error handling for error case with bad data in obidet
This came up for obsid 16502 which had an obidet tstart which
does not appear to be midslew.  This is probably due to some
manual database handling.  While that should probably be fixed
in the database, it probably makes just as much sense to just
use AIPROPS a bit differently (find end of last NMAN directly,
instead of finding the beginning of the last NPNT) if this
happens again.